feat: refactor optuna objective into a separate function
